+++ lib/dartastic_opentelemetry_api.dart
@@ -52,6 +52,7 @@ export 'src/api/otel_api.dart';
 export 'src/api/semantics/lifecycle_state.dart';
 export 'src/api/semantics/navigation_action.dart';
 export 'src/api/semantics/resource_semantics.dart';
+export 'src/api/semantics/resource_values.dart';
 // Semantics
 export 'src/api/semantics/semantics.dart';
 export 'src/api/semantics/ui_semantics.dart';